refactor: update import paths from db to src.db
